// src/app/models/usuario.ts
import { Rol } from './rol';

export class Usuario {
  idUsuario: number = 0;
  nombre: string = '';
  apellido: string = '';
  correo: string = '';
  contrasena: string = '';
  dni: string = '';
  especialidad: string = '';
  credencial: string = '';
  enabled: boolean = true;
  idRol: number = 0;
}